Fix #7781: Avoid using nested package definition for Type
Had to remove `sealed` from `Type`
